muse update 2
I've found that the root note gets changed often enough that it
shouldn't be necessary to prepend 'k' whenever we want to alter it. I've
made it so that changing the root note is the default behavior when
feeding [muse] a list of numbers. Now, when 'k' is prepended, it's meant
to suggest that the key is implicit, or stays the same.

In short, muse_list and muse_key have switched in behavior, and
muse_skip has been removed completely.