mantr.pd & mantouch.pd
created two additional mantissa display objects.

mantr is a version of mantissa.pd that also includes an outlet that
returns the value as a 23-bit integer.

mantouch is a more interactive approach, where the user manually
switches on/off the toggles and the value is updated and outputed with
each change.