chrd and rand update
Made it so that chrd no longer automatically adds an octave to the scale. Octaves now need to be stated explicitly.

Added a special case scenario to rand in the event that you want a list of numbers but only consisting of two numbers, in which case you write "x or y"
ex: [rand 3 or 6]