rand update
changed rand so that the min and max values are ordered more appropriately when there are two or zero creation arguments. Also created a new rand help file that is easier to read.

changed the location of the musl license to be inside of radx.c instead of its own separate file

m_pd.h has been updated to version 49 of pd-vanilla
makefile.pdlibbuilder has been updated to version 0.5.1