rollback changes to before adding pd-lib-builder
Have to do this again one more time because I messed it up the first
time. To reiterate, back when pd-lib-builder was added to the repo, it
did stuff to the commit history that didn't sit well with me, so I'm
rolling back to before that happened so I can add it as a submodule.