adding my fork of game-music-emu as submodule
The game-music-emu fork includes all the changes I've made to the
library in the process of making the gme~ and gmes~ externals.

I've also added delp, linp, and linp~ abstractions, which are delay,
line, and line~ but with a pause feature added. These are currently
being used in the gme~-help file for playback handling.