forgot to remove hot and rev builds
I'd like to use this commit to do a rundown of the commits lost in the
rollback:

In March-April of 2019:

[pak ], [unpak ], & [is ]
unpak - like unpack but uses 'anything' outlets by default and favors no
error messages and no strict type checking while still leaving those
options open.
pak - unpak's counterpart
is - checks an atom's type and returns 1 or 0 if it matches the type
it's looking for.

x - a trigger object that de-husks 1-atom lists to help with
type-checking

@pak/@unpak - reverse pak/unpak
  outputs the stored values in reverse order

help files now have large labels for the object's name

cupq and cupqb are now set up so that both can have their values set
without outputting them via the middle inlet, and have different step
values between the left and right inlets.

renamed graid to grdnt (short for gradient) and radx to radix

In August-September of 2019:

muse, chrd, and rand have had their "set" message renamed to "@", since
"set" was conflicting with message-boxes also taking "set" messages.

changed the name of freqm and fmbloc to fmod and fmad respectively.

rand update - no-repeat mode is now automatically in effect when floats
are used so there's no need to turn nop on prior to sending floats to
rand.

Finally got around to adding a gme help file. Also included a gme-mask
abstraction for muting and un-muting specific channels, and a gme-test
patch that tests out the multi-gme setup.

Another rand update: Sep 12th
There's been a rather large update to rand. The code has been simplified
with better explanations of what's going on, and its behavior has been
altered slightly to coincide more with its high-precision cousin, rind,
and with pd's native random object.

Before, when rand was given two arguments, they would be interpreted as
the lowest and highest possible outcomes. For example: [rand 4 6] would
have 3 possible outcomes (4,5,6). Now to get the same results, you would
instead write [rand 7 4] (or [rand 4 7]). The benefit of this style is
that 7-4 = 3 (the range), so no additional logic needs to be performed
to get the correct range.

The min and range values are also no longer ints. If we gave rand the
arguments 1.9 3.1, it would mostly result in 2's with the occasional 1's
and 3's.

rand and rind also now accept lists for fast assignment of min and max
values.