Added Blunt! library
Overrides f, i, and many binops objects to add a feature where
if you type an exclamation point after the number, it will output the number or calculation with a loadbang function.
Additionally, it allows for binops objects to allow 2 arguments,
and if an exclamation point is at the end of the 2nd argument, the calculation will be outputted.

This only seems to work for more recent versions of Pd.
At some point, Pd started allowing for the overriding of built-in objects.
Currently, Pd-l2ork/Purr-Data do not seem to support this feature.