muse: major update
All of the list types have the option to add a starting index.
For implicit, the syntax is [!$1 ... (
For explicit, the syntax is [@$1 ... (
For neutral,  the syntax is [#$1 ... (

The @ and ! symbols have also replaced x and i messages.

^ and v are no longer + and - aliases.
Instead, the tilde is an alias for minus, since it's the only
operator that can be misinterpreted as a negative sign.

You can now sharpen or flatten a note without altering
the scale, using this syntax: [$1+$2 (
with $1 being the scale reference,
and $2 being any addtional semitones to add to or subtract from it.