ffplay and gme now use a common header
These objects shared a lot of the same code. Using a commmon
header means that there will be less redundant code.

Added a new abstraction called [tick], which is a clicking metronome
that also averages out the delays, making it relatively easy to
determine a song's tempo.

Made a small change in game-music-emu. The FDS blip buffer is
now a lower resolution.